Added tests for poptorch user guide examples
Summary:
The following scripts were also identified to be failing:
- device_iterations.py
- ~~half_float_ops.py~~
- ~~pipeline_simple.py~~

Reviewers: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb, nicholasw

Reviewed By: #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, #pytorch, anthonyb, nicholasw

Subscribers: nicholasw, anthonyb

Maniphest Tasks: T28429

Differential Revision: https://phabricator.sourcevertex.net/D35920

+++ docs/user_guide/half_float_ops.py
@@ -78,4 +78,4 @@ poptorch_model = poptorch.inferenceModel(native_model)
 assert poptorch_model(float16_tensor).dtype == torch.float16
 
 poptorch_model = poptorch.inferenceModel(native_model)
-assert poptorch_model(float32_tensor).dtype == torch.float32
+assert poptorch_model(float32_tensor).dtype == torch.float16
